backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
requests
python-dotenv
middleware
joblib
scikit-learn